refactor(Service): rename from TitleSearchStrategy to SearchVideoByTitleService
